% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{agree_strongly_agree2}
\alias{agree_strongly_agree2}
\title{Agree/Strongly agree}
\usage{
agree_strongly_agree2(data, name)
}
\arguments{
\item{data}{the data}

\item{name}{a string - the question to get the percentage for}
}
\value{
a string
}
\description{
Gets the percent that agree and strongly agree with different variable names
}
