% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coalesce_by_column}
\alias{coalesce_by_column}
\title{Coalesce everything}
\usage{
coalesce_by_column(df)
}
\arguments{
\item{df}{the dataframe}
}
\value{
the dataframe coalesced
}
\description{
Takes all columns and splices them into dots for combination
}
\examples{
df <- data.frame(
  a = c(1, 2, 3),
  b = c(NA, NA, NA),
  id = c("xxx", "xxx", "xxx")
)
df \%>\%
  dplyr::group_by(id) \%>\%
  dplyr::summarise_all(TeachingLab::coalesce_by_column)
}
