% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colorize}
\alias{colorize}
\title{Set Color in rmd with HTML or LaTex}
\usage{
colorize(x, color)
}
\arguments{
\item{x}{the text to be colored}

\item{color}{the color of the text}
}
\value{
Returns a string with code to render the correctly colored text
}
\description{
Creates code to generate a color in x text with color, color.
}
