% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{conditional_slice_sample}
\alias{conditional_slice_sample}
\title{Conditionally slice}
\usage{
conditional_slice_sample(data, max)
}
\arguments{
\item{data}{the data}

\item{max}{a max length to slice the data for}
}
\value{
a string
}
\description{
Conditionally reduces the data
}
