% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{conditionally}
\alias{conditionally}
\title{Conditionally Perform Function}
\usage{
conditionally(fun)
}
\arguments{
\item{fun}{the function to wrap}
}
\description{
Wraps a function and conditionally performs it given certain arguments
}
\examples{
library(dplyr)
cond_filter <- conditionally(filter)
cond_select <- conditionally(select)
}
