% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{dashboard_ipg_plot_ts}
\alias{dashboard_ipg_plot_ts}
\title{IPG Forms Grapher}
\usage{
dashboard_ipg_plot_ts(
  data,
  name,
  wrap = 60,
  sizing = 1,
  split = F,
  numeric = F,
  factor_level = NULL
)
}
\arguments{
\item{data}{the data}

\item{name}{the column name for the data frame to focus on}

\item{wrap}{passes to str_wrap for the title}

\item{sizing}{the base text size multiplier}

\item{split}{if it is sequenced by commas split it}

\item{numeric}{if it is numeric reorder the factors}

\item{factor_level}{option for factor levels}
}
\value{
a ggplot object
}
\description{
Creates a graph specifically for IPG Forms data
}
