% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{fake_bar_graph_create}
\alias{fake_bar_graph_create}
\title{FAKE Knowledge Assessment Graph Summary}
\usage{
fake_bar_graph_create(
  title,
  fake_data_fun = "p_and_n",
  custom_n_range = 0,
  custom_n = NULL,
  custom_x_axis_labels = NULL,
  custom_p_1 = NULL,
  custom_p_2 = NULL,
  randomness = 0,
  multiple_labels = NULL,
  know_graph = F
)
}
\arguments{
\item{title}{the title for the knowledge assessment to make plot for}

\item{fake_data_fun}{p_and_n, p_and_n_split}

\item{custom_n_range}{custom n range the number to add/subtract for random number generation}

\item{custom_n}{custom n, one number}

\item{custom_x_axis_labels}{custom x-axis labels}

\item{custom_p_1}{custom percentage 1, vector with 1: initial percentage, and 2: amount to set maximum increase}

\item{custom_p_2}{custom percentage 2, vector with 1: initial percentage, and 2: amount to set maximum increase}

\item{randomness}{multiplier for amount of randomness in custom percentages, default 0}

\item{multiple_labels}{a special label maker to add group labels in bar charts}

\item{know_graph}{if it is a knowledge assessments graph then add \% correct to title}
}
\value{
a ggplot object
}
\description{
Creates a bar graph specifically for Knowledge Assessments for SXSW Report
}
