% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{fake_line_graph_create}
\alias{fake_line_graph_create}
\title{FAKE Knowledge Assessment Line Graph Summary}
\usage{
fake_line_graph_create(
  title,
  fake_data_fun = "time_data",
  labels = c("Test 1", "Test 2"),
  x_axis = "yearly",
  y_axis_label = NULL,
  lines = 2,
  custom_n = NULL,
  custom_n_range = 0
)
}
\arguments{
\item{title}{the title for the plot}

\item{fake_data_fun}{time_data}

\item{labels}{The labels for the lines}

\item{x_axis}{the labels for the x-axis: yearly or pre-pl-post-pl}

\item{y_axis_label}{A custom y axis label}

\item{lines}{number of lines to create, either 1 or 2}

\item{custom_n}{custom n, one number}

\item{custom_n_range}{custom n range the number to add/subtract for random number generation}
}
\value{
a ggplot object
}
\description{
Creates a line graph specifically for Knowledge assessments for SXSW Report
}
