% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{fetch_survey_2}
\alias{fetch_survey_2}
\title{Survey Standardization}
\usage{
fetch_survey_2(id, name)
}
\arguments{
\item{id}{the SurveyMonkey id of the survey}

\item{name}{the name of the survey}
}
\description{
Fetches the knowledge assessment survey specified and standardizes it
}
