% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_fetch.R
\name{get_coaching_feedback}
\alias{get_coaching_feedback}
\title{Coaching Participant Feedback Data}
\usage{
get_coaching_feedback(update = F)
}
\arguments{
\item{update}{FALSE, whether or not to pull the updated version}
}
\value{
Returns a tibble
}
\description{
Gets data from Coaching Participant Feedback
}
