% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_fetch.R
\name{get_course_survey}
\alias{get_course_survey}
\title{End of Course Dashboard Data}
\usage{
get_course_survey(loc = "data/course_surveymonkey.rds")
}
\arguments{
\item{loc}{the location of the end of course data}
}
\value{
Returns a tibble
}
\description{
Gets dashboard data by reading it in from data folder
}
