% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets_fetch.R
\name{get_current_partner_sites}
\alias{get_current_partner_sites}
\title{Ongoing list of partner sites}
\usage{
get_current_partner_sites(update = FALSE, year = "22_23")
}
\arguments{
\item{update}{FALSE, whether or not to pull the updated version}

\item{year}{the year to get the site for}
}
\value{
Returns a tibble of sites
}
\description{
Gets data from Google Sheet of partner sites
}
