% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_surveymonkey_funcs.R
\name{get_family_survey}
\alias{get_family_survey}
\title{Family Survey Data}
\usage{
get_family_survey(update = FALSE)
}
\arguments{
\item{update}{FALSE, whether or not to pull the updated version}
}
\value{
Returns a tibble
}
\description{
Gets data from Family Survey
}
