% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_fetch.R
\name{get_followup_educator}
\alias{get_followup_educator}
\title{Follow Up Educator Survey Data}
\usage{
get_followup_educator(update = FALSE)
}
\arguments{
\item{update}{FALSE, whether or not to pull the updated version}
}
\value{
Returns a tibble
}
\description{
Gets data from the Follow Up Educator Survey
}
