% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_fetch.R
\name{get_knowledge_assessments}
\alias{get_knowledge_assessments}
\title{Knowledge Assessments Update}
\usage{
get_knowledge_assessments(update = FALSE, year = "22_23")
}
\arguments{
\item{update}{FALSE, optional updating}

\item{year}{"21_22" or "22_23"}
}
\value{
A tibble
}
\description{
Get the knowledge assessments survey
}
