% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets_fetch.R
\name{get_lesson_analysis}
\alias{get_lesson_analysis}
\title{Lesson Plan Analysis Data}
\usage{
get_lesson_analysis(update = FALSE)
}
\arguments{
\item{update}{FALSE}
}
\value{
Returns a tibble
}
\description{
Gets data from Lesson Plan Analysis forms
}
