% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_fetch.R
\name{get_ongoing_coaching}
\alias{get_ongoing_coaching}
\title{Ongoing Coaching Feedback Survey Data}
\usage{
get_ongoing_coaching(update = FALSE)
}
\arguments{
\item{update}{FALSE, whether or not to pull the updated version}
}
\value{
Returns a tibble
}
\description{
Gets data from the Ongoing Coaching Feedback Survey
}
