% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_fetch.R
\name{get_ongoing_coaching_feedback}
\alias{get_ongoing_coaching_feedback}
\title{Coaching Participant Feedback Data}
\usage{
get_ongoing_coaching_feedback(update = FALSE, year = "21_22")
}
\arguments{
\item{update}{FALSE, whether or not to pull the updated version}

\item{year}{"21_22" or "22_23"}
}
\value{
Returns a tibble
}
\description{
Gets data from Coaching Participant Feedback
}
