% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_fetch.R
\name{get_participant_feedback}
\alias{get_participant_feedback}
\title{Participant Feedback Data}
\usage{
get_participant_feedback(update = FALSE, start_date = as.Date("2022-07-01"))
}
\arguments{
\item{update}{FALSE, whether or not to update}

\item{start_date}{the initial date to get data from, defaults to start of data collection}
}
\value{
Returns a tibble
}
\description{
Gets all participant feedback qualtrics data
}
