% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_season}
\alias{get_season}
\title{Get Season}
\usage{
get_season(date)
}
\arguments{
\item{date}{the date}
}
\value{
the season
}
\description{
Takes a date and finds the season
}
\examples{
get_season(as.POSIXct("2016-01-01 12:00:00"))
}
