% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets_fetch.R
\name{get_student_scores_mississippi2}
\alias{get_student_scores_mississippi2}
\title{Student Scores Mississippi Round 2}
\usage{
get_student_scores_mississippi2(update = FALSE)
}
\arguments{
\item{update}{FALSE, optional updating}
}
\value{
A tibble
}
\description{
Get student scores for follow up mississippi data
}
