% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_fetch.R
\name{get_student_survey}
\alias{get_student_survey}
\title{Student Survey Data}
\usage{
get_student_survey(update = FALSE, year = "22_23")
}
\arguments{
\item{update}{FALSE whether or not to update the data}

\item{year}{"21_22" or "22_23"}
}
\value{
Returns a tibble
}
\description{
Gets data from Student Survey
}
