% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_fetch.R
\name{get_student_work}
\alias{get_student_work}
\title{Student Work Data}
\usage{
get_student_work(update = FALSE, year = "22_23")
}
\arguments{
\item{update}{FALSE, whether or not to pull the updated version}

\item{year}{"21_22" or "22_23"}
}
\value{
Returns a tibble
}
\description{
Gets metadata about student work files
}
