% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets_fetch.R
\name{get_student_work_grades}
\alias{get_student_work_grades}
\title{Ongoing list of partner sites}
\usage{
get_student_work_grades(year = "22_23")
}
\arguments{
\item{year}{the year to get the site for}
}
\value{
Returns a tibble of sites
}
\description{
Gets data from Google Sheet of partner sites
}
