% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Find elements x not in a vector y}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{A vector of what shouldn't exist}

\item{y}{A vector to check against}
}
\value{
Returns elements not in vector
}
\description{
Find elements x not in a vector y
}
