% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_makers.R
\name{gt_arrow}
\alias{gt_arrow}
\title{Arrow maker for gt table}
\usage{
gt_arrow(data, colors = c("#800000", "#98AFC7"), column_one, column_two)
}
\arguments{
\item{data}{the gt table to make arrows for}

\item{colors}{the color style to apply}

\item{column_one}{The first column to compare}

\item{column_two}{The second column to compare}
}
\value{
a colored gt table
}
\description{
Makes an html column for a gt table
}
\examples{
gt_arrow(data = mtcars, colors = c("red", "blue"), column_one = cyl, column_two = disp)
}
