% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_makers.R
\name{gt_know_assess}
\alias{gt_know_assess}
\title{Knowledge Assessment Graphs}
\usage{
gt_know_assess(data, know_assess)
}
\arguments{
\item{data}{the data}

\item{know_assess}{the knowledge assessment to make a table for}
}
\value{
a gt table
}
\description{
Creates a graph specifically for Knowledge Assessments in mid year reports
}
