% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gt_theme_tl}
\alias{gt_theme_tl}
\title{Create Teaching Lab theme to a gt table}
\usage{
gt_theme_tl(data, all_caps = F, ...)
}
\arguments{
\item{data}{An existing gt table object}

\item{...}{Optional additional arguments to gt::table_options()}
}
\value{
Returns a tibble
}
\description{
Create Teaching Lab theme to a gt table
}
