% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependencies.R
\name{html_dependency_jquery_stickytableheaders}
\alias{html_dependency_jquery_stickytableheaders}
\title{create an html dependency for jquery-stickytableheaders}
\usage{
html_dependency_jquery_stickytableheaders()
}
\description{
create an html dependency for jquery-stickytableheaders
}
