% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependencies.R
\name{html_dependency_magnific_popup}
\alias{html_dependency_magnific_popup}
\title{create an html dependency for Magnific popup}
\usage{
html_dependency_magnific_popup()
}
\description{
create an html dependency for Magnific popup
}
