% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_template.R
\name{html_template}
\alias{html_template}
\title{Shared HTML template function}
\usage{
html_template(
  template_name,
  template_path,
  template_dependencies,
  pandoc_args,
  ...
)
}
\arguments{
\item{template_name}{template file name}

\item{template_path}{template location}

\item{template_dependencies}{js and css dependencies}

\item{pandoc_args}{pandoc additional args}

\item{...}{additional arguments to pass to pkgdown args}
}
\description{
Shared HTML template function
}
