% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{html_wrap}
\alias{html_wrap}
\title{HTML Text Wrapping}
\usage{
html_wrap(string, n = 40)
}
\arguments{
\item{string}{the string}

\item{n}{the width of the string before a \if{html}{\out{<br>}} tag}
}
\value{
the same string with \if{html}{\out{<br>}} inserted at the requested interval
}
\description{
Takes a string and inserts \if{html}{\out{<br>}} at the requested intervals
}
\examples{
html_wrap("a random string that has about 40 characters in it")
}
