% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{id_maker}
\alias{id_maker}
\title{ID Maker}
\usage{
id_maker(initials, birthday)
}
\arguments{
\item{initials}{the initials}

\item{birthday}{the birthday}
}
\description{
takes initials and id column and makes an ID
}
