% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{ipg_plot}
\alias{ipg_plot}
\title{IPG Forms Grapher}
\usage{
ipg_plot(
  data,
  name,
  save_name,
  height = 5,
  width = 8.5,
  wrap = 60,
  sizing = 1,
  dpi = 300,
  split = F,
  numeric = F,
  factor_level = NULL,
  save = FALSE
)
}
\arguments{
\item{data}{the data}

\item{name}{the column name for the data frame to focus on}

\item{save_name}{the name to save}

\item{height}{height}

\item{width}{width}

\item{wrap}{passes to str_wrap for the title}

\item{sizing}{the base text size multiplier}

\item{dpi}{the dpi to save with}

\item{split}{if it is sequenced by commas split it}

\item{numeric}{if it is numeric reorder the factors}

\item{factor_level}{option for factor levels}

\item{save}{FALSE}
}
\value{
a ggplot
}
\description{
Creates a graph specifically for IPG Forms data
}
