% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{know_assess_summary}
\alias{know_assess_summary}
\title{Knowledge Assessment Graph Summary}
\usage{
know_assess_summary(data, know_assess)
}
\arguments{
\item{data}{the data}

\item{know_assess}{the knowledge assessment to make a table for}
}
\value{
a gt table
}
\description{
Creates a graph specifically for Knowledge Assessments in mid year reports
}
