% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{knowledge_assess_detailed_score}
\alias{knowledge_assess_detailed_score}
\title{Knowledge Assessments Question Scoring 2022-2023}
\usage{
knowledge_assess_detailed_score(survey_id, survey_name)
}
\arguments{
\item{survey_id}{Qualtrics survey id}

\item{survey_name}{Name of survey - should match Qualtrics name}
}
\value{
a dataframe of columns: prepost, site, know_assess, date, question, score, question2, answer, max_score
}
\description{
function to grade and output each question per respondent in a knowledge assessments survey
}
