% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{knowledge_assess_select_score}
\alias{knowledge_assess_select_score}
\title{Knowledge Assessments Scoring 2022-2023}
\usage{
knowledge_assess_select_score(survey_id, survey_name)
}
\arguments{
\item{survey_id}{Qualtrics survey id}

\item{survey_name}{Name of survey - should match Qualtrics name}
}
\value{
a dataframe of columns: id, percent, prepost, site, know_assess, date
}
\description{
function to grade and output specific set of data points for knowledge assessments
}
