% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mutate_cond}
\alias{mutate_cond}
\title{Conditionally mutate}
\usage{
mutate_cond(.data, condition, ..., envir = parent.frame())
}
\arguments{
\item{.data}{the data}

\item{condition}{a specific condition to test for}

\item{...}{additional arguments for mutate}

\item{envir}{parent.frame()}
}
\value{
a string
}
\description{
Conditionally changes the data
}
