% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{neg_cond_filter}
\alias{neg_cond_filter}
\title{Negative Conditional Filter}
\usage{
neg_cond_filter(not_this, is_this, dat_filter)
}
\arguments{
\item{not_this}{If value is not this}

\item{is_this}{Filter for this}

\item{dat_filter}{Data column to filter}
}
\value{
filtered dataframe
}
\description{
Conditionally filters value given that it is not the first parameter, for use in shiny apps
}
