% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_generating.R
\name{ongoing_coaching_feedback_graph}
\alias{ongoing_coaching_feedback_graph}
\title{Ongoing Coaching feedback graph dependent on race and content area}
\usage{
ongoing_coaching_feedback_graph(
  race_filter = "All",
  content_area_filter = "All"
)
}
\arguments{
\item{race_filter}{the filter to apply to the race column, one of "All", "White", "Black or African American", "Asian", "Hispanic/Latino", "More than one race", or "Other"}

\item{content_area_filter}{the content area to filter for, one otf "All", "ELA", "Math", or "Other"}
}
\value{
prints a ggplot object
}
\description{
Returns a barchart for selections in the relevant questions of the end of coaching survey
}
