% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{p_and_n_split}
\alias{p_and_n_split}
\title{Fake p1_1, p1_2, p2_1, p2_2, n1, n2 data}
\usage{
p_and_n_split(
  p1_1_range = c(30:70),
  p1_2_range = c(70:100),
  p2_1_range = c(80:95),
  p2_2_range = c(85:100),
  n1_range = stats::rnorm(1, mean = 80, sd = 10),
  n2_range = stats::rnorm(1, mean = 50, sd = 10)
)
}
\arguments{
\item{p1_1_range}{the lower range of values to sample for p1}

\item{p1_2_range}{the upper range of values to sample for p1}

\item{p2_1_range}{the lower range of values to sample for p2}

\item{p2_2_range}{the upper range of values to sample for p2}

\item{n1_range}{the range of values to sample for n1}

\item{n2_range}{the range of values to sample for n2}
}
\value{
a randomised dataset
}
\description{
Creates fake data for knowledge assessments in SXSW report
}
