% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_generating.R
\name{partner_file_remove}
\alias{partner_file_remove}
\title{Remove Image Files Render R Markdown}
\usage{
partner_file_remove(partner)
}
\arguments{
\item{partner}{The partner to render a report for}
}
\description{
file remove and render
}
