% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{password_generator}
\alias{password_generator}
\title{Password Generator}
\usage{
password_generator(length = 8)
}
\arguments{
\item{length}{the length of the password}
}
\value{
a string
}
\description{
Creates a password of n length
}
