% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{percent_agree}
\alias{percent_agree}
\title{Percent Agree/Strongly agree}
\usage{
percent_agree(agree_col)
}
\arguments{
\item{agree_col}{the vector or column to summarise}
}
\value{
integer
}
\description{
Calculates percent that are 4, 5, agree, or strongly agree
}
