% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{quote_viz}
\alias{quote_viz}
\title{Quote Visualization}
\usage{
quote_viz(data, text_col, viz_type = c("ggplot", "gt"))
}
\arguments{
\item{data}{the dataframe}

\item{text_col}{the text to visualize}

\item{viz_type}{ggplot or gt visualization}
}
\value{
a ggplot/gt that visualizes text
}
\description{
takes a dataframe and makes a gt table or ggplot that shows a quote
}
\examples{
quote_viz(data = iris, text_col = Species, viz_type = "ggplot")
}
