% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rand_vect}
\alias{rand_vect}
\title{Random Vector}
\usage{
rand_vect(N, M, sd = 1, pos.only = TRUE)
}
\arguments{
\item{N}{number of scalars to create}

\item{M}{number to sum to}

\item{sd}{Standard Deviation}

\item{pos.only}{T}
}
\description{
Random Vector
}
