% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{relabel_qualtrics_df}
\alias{relabel_qualtrics_df}
\title{Qualtrics Data Frame Relabel}
\usage{
relabel_qualtrics_df(df, switch = TRUE)
}
\arguments{
\item{df}{the relevant dataframe}

\item{switch}{T/F whether or not to switch the names}
}
\value{
a data.frame
}
\description{
Switches the labels and column names of a data frame, or just changes
the column names to the labels
}
\examples{
\dontrun{
qualtrics_survey |>
  relabel_qualtrics_df()
}
}
