% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{round_even}
\alias{round_even}
\title{Round to nearest even number}
\usage{
round_even(x)
}
\arguments{
\item{x}{the number to round}
}
\value{
an integer
}
\description{
Takes a whole number and "rounds" it to the nearest even number.
}
\examples{
round_even(17)
}
