% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runif_round}
\alias{runif_round}
\title{Generate one random number from a min and max}
\usage{
runif_round(min, max)
}
\arguments{
\item{min}{the minimum number allowed}

\item{max}{the maximum number allowed}
}
\value{
a single integer
}
\description{
Generate one random number from a min and max
}
