% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_assessments_formatting.R
\name{save_processed_data}
\alias{save_processed_data}
\title{Save Processed Knowledge Assessments Data}
\usage{
save_processed_data(data, q_and_a, correct, save_name, question_html_wrap = 45)
}
\arguments{
\item{data}{the dataframe to be evaluated}

\item{q_and_a}{a dataframe of questions and answers}

\item{correct}{the correct answers}

\item{save_name}{a folder for the plot ready data to be saved}

\item{question_html_wrap}{number of characters before <br> insertion in question}
}
\value{
a plot ready dataframe
}
\description{
A function to use on already loaded data
}
