% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scale_dollar_labels}
\alias{scale_dollar_labels}
\title{Five thirty-eight style formatter for currency}
\usage{
scale_dollar_labels(labels)
}
\arguments{
\item{labels}{vector of labels}
}
\value{
formatted percent labels
}
\description{
Five thirty-eight style formatter for currency
}
