% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scale_percent_labels}
\alias{scale_percent_labels}
\title{Five thirty-eight style formatter for percentages}
\usage{
scale_percent_labels(labels)
}
\arguments{
\item{labels}{vector of labels}
}
\value{
formatted percent labels
}
\description{
Five thirty-eight style formatter for percentages
}
