% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scale_ratio_labels}
\alias{scale_ratio_labels}
\title{Five thirty-eight style formatter for Ratios}
\usage{
scale_ratio_labels(labels)
}
\arguments{
\item{labels}{vector of labels}
}
\value{
formatted ratio labels
}
\description{
Five thirty-eight style formatter for Ratios
}
