% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{score_compare_plot}
\alias{score_compare_plot}
\title{Graph maker for score comparisons}
\usage{
score_compare_plot(
  data,
  question,
  order,
  prepost,
  score,
  split_variable,
  title
)
}
\arguments{
\item{data}{the data to use}

\item{question}{the question column to use for comparison}

\item{order}{the order to keep the questions in}

\item{prepost}{The column to compare before and after}

\item{score}{the scores to compare}

\item{split_variable}{the text detected in the variable to create segments with (from prepost)}

\item{title}{the title of the gt}
}
\value{
a colored gt table
}
\description{
Makes a ggplot comparison table
}
\examples{
\dontrun{
score_compare_plot(data, question, prepost, score)
}
}
