% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{score_question_improved}
\alias{score_question_improved}
\title{Scores pre and post and percent improved/sustained}
\usage{
score_question_improved(
  data,
  question_pre,
  question_post,
  coding,
  middle_value
)
}
\arguments{
\item{data}{the dataframe to be analyzed}

\item{question_pre}{the initial column to be selected}

\item{question_post}{the comparison column to be selected}

\item{coding}{the coding to check for}

\item{middle_value}{the middle value to check for when calculating scores}
}
\value{
Returns a dataframe with the percent, correct, number of non-na responses, the question itself, and the percent that sustained/improved
}
\description{
Calculate percentage of a pre question (column) and post question (column) in data (data)
that is in the right answer (coding).
}
