% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{score_question_mindsets}
\alias{score_question_mindsets}
\title{Mindsets scoring}
\usage{
score_question_mindsets(
  data,
  question_pre,
  question_post,
  coding,
  na_remove = F,
  likert = c(5, 6)
)
}
\arguments{
\item{data}{the dataframe to be analyzed}

\item{question_pre}{the initial column to be selected}

\item{question_post}{the comparison column to be selected}

\item{coding}{the coding to check for}

\item{na_remove}{whether or not to drop NAs at the start of the evaluation}

\item{likert}{whether or not the scale is likert with 5 points or 6}
}
\value{
Returns a dataframe with the percent, correct, number of non-na responses, the question itself, and the percent that sustained/improved
}
\description{
Calculate percentage correct for mindsets & expectations
}
