% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_generating.R
\name{select_useful_text}
\alias{select_useful_text}
\title{Useful text selector}
\usage{
select_useful_text(data, text_col, name, n = 15)
}
\arguments{
\item{data}{the data to filter the text for}

\item{text_col}{the text column to select}

\item{name}{the new name for the text column}

\item{n}{15 observations to return by default}
}
\value{
returns a single tibble column
}
\description{
Gets all text over a specified nps and tibble
}
