% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{selection_sum}
\alias{selection_sum}
\title{Summarise a selection}
\usage{
selection_sum(data, select, group = NULL)
}
\arguments{
\item{data}{the data}

\item{select}{the data to select}

\item{group}{the data to group by}
}
\value{
a tibble of the selected data's percent correct
}
\description{
finds the average sum of all selected columns
}
