% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{session_agree_plot}
\alias{session_agree_plot}
\title{Dashboard Agree Type \% Plot}
\usage{
session_agree_plot(data)
}
\arguments{
\item{data}{the data to be input}
}
\value{
Returns a ggplot
}
\description{
Creates a plot for session survey data that shows \% in each Likert category
}
