% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{session_agree_plot_ts}
\alias{session_agree_plot_ts}
\title{Dashboard Time Series Plot}
\usage{
session_agree_plot_ts(data, scale = "1 month")
}
\arguments{
\item{data}{the data to be input}

\item{scale}{the date scale to use for the plot}
}
\value{
Returns a ggplot
}
\description{
Creates a plot for session survey data with an adjustable time series component
}
