% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{single_sort_view}
\alias{single_sort_view}
\title{Unique, sort, view}
\usage{
single_sort_view(vector)
}
\arguments{
\item{vector}{the vector to view}
}
\value{
a string
}
\description{
Gets all unique elements of a vector in a sorted view
}
\examples{
# example code
 c("stuff", "and", "more") |> single_sort_view()
}
