% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{site_condense}
\alias{site_condense}
\title{Site Replacement}
\usage{
site_condense(data)
}
\arguments{
\item{data}{the data to perform replacement on}
}
\value{
a string
}
\description{
Replaces individual sites with greater groupings
}
