% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{string_replace}
\alias{string_replace}
\title{String Replace}
\usage{
string_replace(string, string_detect, string_replace, print = FALSE)
}
\arguments{
\item{string}{the string to look within}

\item{string_detect}{the string to detect}

\item{string_replace}{the string to replace detected string with}

\item{print}{FALSE whether or not to print the string}
}
\value{
a string
}
\description{
Detects a specific string pattern and replaces entire string with it
}
