% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{strip_html}
\alias{strip_html}
\title{Remove HTML}
\usage{
strip_html(string)
}
\arguments{
\item{string}{a string}
}
\value{
a string without the html
}
\description{
Takes a string and removes html using rvest
}
