% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_makers.R
\name{table_maker}
\alias{table_maker}
\title{General Table Maker}
\usage{
table_maker(
  data,
  column_names,
  title,
  spanner,
  n1,
  n2,
  rows_positive,
  rows_negative,
  improve_col,
  bottom_row
)
}
\arguments{
\item{data}{the dataframe to be analyzed}

\item{column_names}{the names of the columns}

\item{title}{The tables title}

\item{spanner}{The gt spanner}

\item{n1}{the n for fall}

\item{n2}{the n for spring}

\item{rows_positive}{the positive rows}

\item{rows_negative}{the negative_rows}

\item{improve_col}{name for the improve/sustain column}

\item{bottom_row}{the bottom row which has "n = " instead of data}
}
\value{
Returns an unsaved gt table
}
\description{
Makes a gt table for mindsets question set
}
