% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{temp_save}
\alias{temp_save}
\title{Temporary Image Save and Return}
\usage{
temp_save(img)
}
\arguments{
\item{img}{the object to save as an image}
}
\value{
a png image
}
\description{
Creates a temporary save location for an image and return it with \code{knitr::include_graphics}
}
