% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{tl_likert}
\alias{tl_likert}
\title{TL Themed Stacked Bar Chart}
\usage{
tl_likert(data, title = "\% Selected", string_remove = NULL, string_wrap = 80)
}
\arguments{
\item{data}{the data for the plotter to use, should include all columns of interest}

\item{title}{the title for the plot}

\item{string_remove}{NULL by default, provides an optional string removal}

\item{string_wrap}{80 by default, the amount by which to wrap y axis text}
}
\value{
a ggplot object
}
\description{
Automatically scaled stacked bar chart with TL theming
}
