% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{tl_pal}
\alias{tl_pal}
\title{A muted, qualitative color palette}
\usage{
tl_pal()
}
\description{
A muted, qualitative color palette
}
\examples{
\dontrun{
library(scales)
scales::show_col(tl_pal_blue()(9))
}
}
