% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{tl_palette}
\alias{tl_palette}
\title{Teaching Lab Color Palette Maker}
\usage{
tl_palette(
  color = c("blue", "orange", "purple", "green", "teal", "tl_colors"),
  theme = c("dark", "light"),
  n
)
}
\arguments{
\item{theme}{if theme is light or dark}

\item{n}{number of colors to generate}

\item{palette}{choice of color palette}
}
\value{
color ramp palette function
}
\description{
Teaching Lab Color Palette Maker
}
