% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{tl_score}
\alias{tl_score}
\title{Grade Data}
\usage{
tl_score(data, answer)
}
\arguments{
\item{data}{the data}

\item{answer}{the answer}
}
\value{
a percentage of correct
}
\description{
function for grading in general
}
