% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{tl_score_numeric}
\alias{tl_score_numeric}
\title{IPG Scoring Numeric}
\usage{
tl_score_numeric(x)
}
\arguments{
\item{x}{the numbers to apply to}
}
\value{
a rounded average
}
\description{
function for grading in general
}
