% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{tl_score_percent}
\alias{tl_score_percent}
\title{Get the percent of \code{data} in \code{answer}}
\usage{
tl_score_percent(data, answer)
}
\arguments{
\item{data}{the data}

\item{answer}{the answer}
}
\value{
a vector percent of correct
}
\description{
function for general grading of percentages
}
