% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_makers.R
\name{tl_summary_table}
\alias{tl_summary_table}
\title{Teaching Lab Summary Tables}
\usage{
tl_summary_table(
  data,
  save = F,
  summarise = T,
  grouping = NULL,
  n_size = NULL,
  n_size_single = NULL,
  explain = F,
  prepost = T,
  admin = F,
  rename = F
)
}
\arguments{
\item{data}{the data to be input}

\item{save}{F by default, a save path inside current working directory}

\item{summarise}{TRUE by default will summarise all data selected}

\item{grouping}{includes "equitable", "high_expectations", and "crse"}

\item{n_size}{NULL an optional n_size to include}

\item{n_size_single}{NULL an optional single value for n_size}

\item{explain}{add explanation for different percentage levels to the table}

\item{prepost}{F by default for making a prepost dataframe}

\item{admin}{F by default for prepost admin dataframe}

\item{rename}{F by default for making prepost dataframe}
}
\value{
Returns a gt table, unless save in which case it will return a saved file with gtsave
}
\description{
Creates tables for responses to the Educator Diagnostic Survey
}
