% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{tl_wordcloud}
\alias{tl_wordcloud}
\title{Wordcloud Teaching Lab Visualization}
\usage{
tl_wordcloud(
  data,
  text_col,
  colors = c("blue", "orange"),
  n_min = 2,
  size = 20,
  shape = c("circle", "cardioid", "diamond", "square", "triangle-forward",
    "triangle-upright", "pentagon", "star")
)
}
\arguments{
\item{data}{the dataframe}

\item{text_col}{the text to visualize}

\item{colors}{the color gradient}

\item{n_min}{the minimum n to accept}

\item{size}{the wordcloud size}

\item{shape}{the wordcloud shape}
}
\value{
a wordcloud
}
\description{
takes a dataframe and makes a wordcloud
}
\examples{
\dontrun{
tl_wordcloud(data = iris, text_col = Species)
}
}
