% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{coalesce_by_column}
\alias{coalesce_by_column}
\title{Coalesce everything}
\usage{
coalesce_by_column(df)
}
\arguments{
\item{df}{the dataframe}
}
\value{
the dataframe coalesced
}
\description{
Takes all columns and splices them into dots for combination
}
\examples{
df \%>\%
   group_by(A) \%>\%
   summarise_all(coalesce_by_column)
}
