% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{course_quotes}
\alias{course_quotes}
\title{Dashboard End of Course Quotes}
\usage{
course_quotes(data = TeachingLab::get_course_survey(), size = 10, n = 3)
}
\arguments{
\item{data}{the data to be input}

\item{size}{the number of rows of quotes to return}

\item{n}{}
}
\value{
Returns a gt
}
\description{
Creates a gt table for qualitative responses
}
