% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_fetch.R
\name{get_course_survey}
\alias{get_course_survey}
\title{End of Course Dashboard Data}
\usage{
get_course_survey(update = FALSE, year = "22_23")
}
\arguments{
\item{update}{FALSE, optional to update end of course data or not}

\item{year}{"21_22" or "22_23"}
}
\value{
Returns a tibble
}
\description{
Gets dashboard data by reading it in from data folder
}
