% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets_fetch.R
\name{get_current_partner_sites}
\alias{get_current_partner_sites}
\title{Ongoing list of partner sites}
\usage{
get_current_partner_sites(update = FALSE, condense = FALSE)
}
\arguments{
\item{update}{FALSE, whether or not to pull the updated version}

\item{condense}{FALSE whether or not to group sites like rochester, district 11, etc.}
}
\value{
Returns a tibble
}
\description{
Gets data from Google Sheet of partner sites
}
