% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_fetch.R
\name{get_diagnostic_survey}
\alias{get_diagnostic_survey}
\title{Diagnostic Survey Update}
\usage{
get_diagnostic_survey(update = FALSE, year = "22_23")
}
\arguments{
\item{update}{FALSE, optional updating}

\item{year}{"21_22" or "22_23"}
}
\value{
A tibble
}
\description{
Get the diagnostic survey
}
