% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets_fetch.R
\name{get_ipg_forms}
\alias{get_ipg_forms}
\title{IPG Data}
\usage{
get_ipg_forms(update = FALSE)
}
\arguments{
\item{update}{FALSE}
}
\value{
Returns a tibble
}
\description{
Gets data from IPG forms
}
