% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets_fetch.R
\name{get_ipg_forms}
\alias{get_ipg_forms}
\title{IPG Data}
\usage{
get_ipg_forms(update = FALSE, year = "22_23")
}
\arguments{
\item{update}{FALSE}

\item{year}{"21_22" or "22_23"}
}
\value{
Returns a tibble
}
\description{
Gets data from IPG forms
}
