% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monday_fetch.R
\name{get_monday_board}
\alias{get_monday_board}
\title{Get Monday Board Data}
\usage{
get_monday_board(board_id, first_col_name)
}
\arguments{
\item{board_id}{Which monday board id to use when getting information}

\item{first_col_name}{What to name the first column in the dataframe}
}
\value{
Returns a tibble
}
\description{
Gets monday board data as a tibble
}
\examples{
TeachingLab::get_monday_board(board_id = 2208860812, first_col_name = "Facilitator")
}
