% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_fetch.R
\name{get_session_survey}
\alias{get_session_survey}
\title{End of Session Dashboard Data}
\usage{
get_session_survey(update = F)
}
\arguments{
\item{update}{FALSE, whether or not to update}
}
\value{
Returns a tibble
}
\description{
Gets dashboard data by reading it in from data folder
}
