% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fetch.R
\name{get_session_survey}
\alias{get_session_survey}
\title{End of Session Dashboard Data}
\usage{
get_session_survey()
}
\value{
Returns a tibble
}
\description{
Gets dashboard data by reading it in from data folder
}
