% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_fetch.R
\name{get_session_survey}
\alias{get_session_survey}
\title{End of Session Dashboard Data}
\usage{
get_session_survey(loc = "data/session_survey_21_22data.rds")
}
\arguments{
\item{loc}{the location of the end of session data}
}
\value{
Returns a tibble
}
\description{
Gets dashboard data by reading it in from data folder
}
