% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_fetch.R
\name{get_session_survey}
\alias{get_session_survey}
\title{End of Session Dashboard Data}
\usage{
get_session_survey(update = FALSE, year = "22_23")
}
\arguments{
\item{update}{FALSE, whether or not to update}

\item{year}{"21_22" or "22_23"}
}
\value{
Returns a tibble
}
\description{
Gets dashboard data by reading it in from data folder
}
