% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_fetch.R
\name{get_student_scores_mississippi}
\alias{get_student_scores_mississippi}
\title{Student Scores Mississippi}
\usage{
get_student_scores_mississippi(update = F)
}
\arguments{
\item{update}{FALSE, optional updating}
}
\value{
A tibble
}
\description{
Get student scores for mississippi data
}
