% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_makers.R
\name{gt_tl_color}
\alias{gt_tl_color}
\title{General Table Maker}
\usage{
gt_tl_color(gt_table, color, column, scale = 1)
}
\arguments{
\item{gt_table}{the gt table to color}

\item{color}{the color style to apply}

\item{column}{The column to apply the color to}

\item{scale}{the scale to apply to}
}
\value{
a colored gt table
}
\description{
Makes a gt table with teaching lab color style
}
