% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{html_wrap}
\alias{html_wrap}
\title{HTML Text Wrapping}
\usage{
html_wrap(string, interval = 40)
}
\arguments{
\item{string}{the string}

\item{interval}{the width of the string before a <br> tag}
}
\value{
the same string with <br> inserted at the requested interval
}
\description{
Takes a string and inserts <br> at the requested intervals
}
\examples{
html_wrap(string, interval)
}
