% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_makers.R
\name{quote_viz}
\alias{quote_viz}
\title{Quote Visualization}
\usage{
quote_viz(
  data,
  text_col = colnames(data)[1],
  extra_cols = NULL,
  viz_type = "gt",
  custom_highlight = F,
  width = 60,
  title = NULL,
  suppress_warnings = T,
  align = "center",
  ...
)
}
\arguments{
\item{data}{the dataframe}

\item{text_col}{the text to visualize}

\item{extra_cols}{Under development, to add additional columns to table}

\item{viz_type}{ggplot or gt visualization}

\item{custom_highlight}{Whether to provide custom highlight arguments through `highlight =` or auto highlight 3 most frequent words}

\item{width}{The width of the table generated}

\item{title}{the title of the ggplot or gt}

\item{suppress_warnings}{T/F suppression of warnings}

\item{align}{the table alignment: "left", "center", "right"}

\item{...}{Arguments passed onto the gt table}
}
\value{
a ggplot/gt that visualizes text
}
\description{
takes a dataframe and makes a gt table or ggplot that shows a quote
}
\examples{
df <- TeachingLab::survey_monkey
quote_viz(data = df, 
          text_col = `What is the learning from this course that you are most excited about trying out?`, 
          viz_type = "gt",
          title = "Responses from Survey Monkey")
}
