% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{quote_viz}
\alias{quote_viz}
\title{Quote Visualization}
\usage{
quote_viz(
  data,
  text_col,
  viz_type = c("ggplot", "gt"),
  custom_highlight = F,
  width = 60,
  title,
  ...
)
}
\arguments{
\item{data}{the dataframe}

\item{text_col}{the text to visualize}

\item{viz_type}{ggplot or gt visualization}

\item{custom_highlight}{Whether to provide custom highlight arguments through `highlight =` or auto highlight 3 most frequent words}

\item{title}{the title of the ggplot or gt}
}
\value{
a ggplot/gt that visualizes text
}
\description{
takes a dataframe and makes a gt table or ggplot that shows a quote
}
\examples{
quote_viz(data = iris, text_col = Species, viz_type = "ggplot")
}
