% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{quote_viz}
\alias{quote_viz}
\title{Quote Visualization}
\usage{
quote_viz(
  data,
  text_col,
  extra_cols = NULL,
  viz_type = "gt",
  custom_highlight = F,
  width = 60,
  title = NULL,
  suppress_warnings = T,
  align = "center",
  ...
)
}
\arguments{
\item{data}{the dataframe}

\item{text_col}{the text to visualize}

\item{viz_type}{ggplot or gt visualization}

\item{custom_highlight}{Whether to provide custom highlight arguments through `highlight =` or auto highlight 3 most frequent words}

\item{width}{The width of the table generated}

\item{title}{the title of the ggplot or gt}

\item{...}{Arguments passed onto the gt table}
}
\value{
a ggplot/gt that visualizes text
}
\description{
takes a dataframe and makes a gt table or ggplot that shows a quote
}
\examples{
library(TeachingLab)
data("survey_monkey")
quote_viz(data = survey_monkey, 
          text_col = `What is the learning from this course that you are most excited about trying out?`, 
          viz_type = "gt",
          title = "Responses from Survey Monkey")
}
