% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{round2}
\alias{round2}
\title{Round}
\usage{
round2(x, n)
}
\arguments{
\item{x}{the vector to round}

\item{n}{the number of digits to round}
}
\value{
the vector provided rounded
}
\description{
round that actually round up 0.5 as it should be
}
