% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{save_processed_data2}
\alias{save_processed_data2}
\title{Version 2 Knowledge Assessments Reformat}
\usage{
save_processed_data2(
  data,
  q_and_a,
  correct,
  save_name,
  question_html_wrap = 45
)
}
\arguments{
\item{data}{the dataframe to be evaluated}

\item{q_and_a}{a dataframe of questions and answers}

\item{correct}{the correct answers}

\item{save_name}{a folder for the plot ready data to be saved}

\item{question_html_wrap}{number of characters before <br> insertion in question}
}
\value{
a plot ready dataframe
}
\description{
A function to use on already loaded knowledge_assessments data, this one saves to data/mid_year_reports
}
\examples{
save_processed_data2(data = here::here("Dashboards/KnowledgeAssessments/data/unprocessed/ELABootcamp-FoundationalSkillsBootcampSkills(K-2).rds"),
q_and_a = here::here("Dashboards/KnowledgeAssessments/data/questions_and_answers/ela_foundational_skills.rds"),
correct = c("Print concepts",
           "Phonological awareness",
           "Fluency",
           "It prompts students to use context clues and pictures to decode words",
           "Utilize a variety of ongoing assessment data to determine the focus of instruction for small groups",
           "Group students by their ongoing phase of development with regard to the foundational skills"),
save_name = "ela_foundational_skills")
}
