% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{score_question}
\alias{score_question}
\title{Calculate percentage of a question (column) in data (data) that is in the right answer (coding)}
\usage{
score_question(data, question, coding, na_type = c("NA", "NR"))
}
\arguments{
\item{data}{the dataframe to be analyzed}

\item{question}{the column to be selected}

\item{coding}{the coding to check for}

\item{na_type}{the form that NA takes - could be "No Response" as in the Participant Feedback dashboard}
}
\value{
Returns a dataframe with the percent, correct, number of non-na responses, and question itself
}
\description{
Calculate percentage of a question (column) in data (data) that is in the right answer (coding)
}
