% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{score_question_number}
\alias{score_question_number}
\title{A title}
\usage{
score_question_number(
  data,
  question_pre,
  question_post,
  coding,
  likert = c(5, 6)
)
}
\arguments{
\item{data}{the dataframe to be analyzed}

\item{coding}{the coding to check for}

\item{question}{the column to be selected}
}
\value{
Returns a dataframe with the percent, correct, number of non-na responses, and question itself
}
\description{
Calculate percentage of a question (column) in data (data) that is on the positive or numeric side (coding) with a grading twist, where three is worth 1 and 1/2 or 4/5 is worth 2
}
