% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{session_quotes}
\alias{session_quotes}
\title{Dashboard Quotes}
\usage{
session_quotes(data)
}
\arguments{
\item{data}{the data to be input}
}
\value{
Returns a gt
}
\description{
Creates a gt table for qualitative responses
}
