% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeachingLab-package.R
\docType{package}
\name{TeachingLab-package}
\alias{TeachingLab}
\alias{TeachingLab-package}
\title{TeachingLab: Teaching Lab Custom R Development Package}
\description{
Provides templates, palettes, and useful data analysis for Teaching Lab purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/duncangates/teachinglab-duncan}
  \item Report bugs at \url{https://github.com/duncangates/teachinglab-duncan/issues}
}

}
\author{
\strong{Maintainer}: Duncan Gates \email{duncan.gates@teachinglab.org}

}
\keyword{internal}
