% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{tl_score_count}
\alias{tl_score_count}
\title{Grade Data}
\usage{
tl_score_count(data, answer)
}
\arguments{
\item{data}{the data}

\item{answer}{the answer}
}
\value{
a count of correct
}
\description{
function for generally getting the count of a data that is correct
}
